Data stucture fixes
Fixed:  '/eureka_data/esc30/snow_mass' so that the second field was not
a replicate of the first.

Fixed: '/eureka_data/esc30/snow_depth' so that units were in meters.

Fixed: readme changed to correct description of the ESC-30 cross
section.

Updated: H5 file verion to 1.2 to reflect changes in data contents.

Thanks Justin Beckers for the data audit!